# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Joventy.cd — Visa Premium SaaS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  client: "client",
  admin: "admin",
} as const;

export interface User {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  phone?: string | null;
  role: UserRole;
  createdAt: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 8 */
  password: string;
  firstName: string;
  lastName: string;
  phone?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type ApplicationDestination =
  (typeof ApplicationDestination)[keyof typeof ApplicationDestination];

export const ApplicationDestination = {
  usa: "usa",
  dubai: "dubai",
  turkey: "turkey",
  india: "india",
} as const;

export type ApplicationStatus =
  (typeof ApplicationStatus)[keyof typeof ApplicationStatus];

export const ApplicationStatus = {
  draft: "draft",
  submitted: "submitted",
  in_review: "in_review",
  approved: "approved",
  rejected: "rejected",
  appointment_scheduled: "appointment_scheduled",
} as const;

export interface Application {
  id: number;
  userId: number;
  applicantName: string;
  destination: ApplicationDestination;
  visaType: string;
  status: ApplicationStatus;
  appointmentDate?: string | null;
  notes?: string | null;
  adminNotes?: string | null;
  passportNumber?: string | null;
  travelDate?: string | null;
  returnDate?: string | null;
  purpose?: string | null;
  price?: number | null;
  isPaid: boolean;
  createdAt: string;
  updatedAt: string;
  user?: User | null;
  unreadCount: number;
}

export type CreateApplicationInputDestination =
  (typeof CreateApplicationInputDestination)[keyof typeof CreateApplicationInputDestination];

export const CreateApplicationInputDestination = {
  usa: "usa",
  dubai: "dubai",
  turkey: "turkey",
  india: "india",
} as const;

export interface CreateApplicationInput {
  applicantName: string;
  destination: CreateApplicationInputDestination;
  visaType: string;
  passportNumber?: string;
  travelDate?: string;
  returnDate?: string;
  purpose?: string;
  notes?: string;
}

export type UpdateApplicationInputStatus =
  (typeof UpdateApplicationInputStatus)[keyof typeof UpdateApplicationInputStatus];

export const UpdateApplicationInputStatus = {
  draft: "draft",
  submitted: "submitted",
  in_review: "in_review",
  approved: "approved",
  rejected: "rejected",
  appointment_scheduled: "appointment_scheduled",
} as const;

export interface UpdateApplicationInput {
  status?: UpdateApplicationInputStatus;
  appointmentDate?: string | null;
  adminNotes?: string | null;
  price?: number | null;
  isPaid?: boolean;
  notes?: string | null;
}

export interface ApplicationListResponse {
  applications: Application[];
  total: number;
  page: number;
  totalPages: number;
}

export interface Message {
  id: number;
  applicationId: number;
  userId: number;
  content: string;
  isFromAdmin: boolean;
  isRead: boolean;
  createdAt: string;
  senderName: string;
}

export interface SendMessageInput {
  content: string;
}

export type AdminStatsByDestination = { [key: string]: number };

export type AdminStatsByStatus = { [key: string]: number };

export interface AdminStats {
  totalApplications: number;
  pendingReview: number;
  approvedThisMonth: number;
  totalClients: number;
  byDestination: AdminStatsByDestination;
  byStatus: AdminStatsByStatus;
  recentApplications: Application[];
}

export type ListApplicationsParams = {
  status?: ListApplicationsStatus;
  destination?: string;
  page?: number;
  limit?: number;
};

export type ListApplicationsStatus =
  (typeof ListApplicationsStatus)[keyof typeof ListApplicationsStatus];

export const ListApplicationsStatus = {
  draft: "draft",
  submitted: "submitted",
  in_review: "in_review",
  approved: "approved",
  rejected: "rejected",
  appointment_scheduled: "appointment_scheduled",
} as const;