# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Joventy.cd — Visa Premium SaaS API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication & user profile
  - name: applications
    description: Visa application management
  - name: messages
    description: Client-admin chat
  - name: documents
    description: Document uploads
  - name: admin
    description: Admin operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /auth/logout:
    post:
      operationId: logoutUser
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List visa applications (client sees own, admin sees all)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, submitted, in_review, approved, rejected, appointment_scheduled]
        - name: destination
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationListResponse"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Create a new visa application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationInput"
      responses:
        "201":
          description: Application created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get a single application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update application (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplicationInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

  /applications/{id}/messages:
    get:
      operationId: getApplicationMessages
      tags: [messages]
      summary: Get messages for an application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard statistics
      responses:
        "200":
          description: Statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
          nullable: true
        role:
          type: string
          enum: [client, admin]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - createdAt

    RegisterInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
      required:
        - email
        - password
        - firstName
        - lastName

    LoginInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token

    Application:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        applicantName:
          type: string
        destination:
          type: string
          enum: [usa, dubai, turkey, india]
        visaType:
          type: string
        status:
          type: string
          enum: [draft, submitted, in_review, approved, rejected, appointment_scheduled]
        appointmentDate:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        adminNotes:
          type: string
          nullable: true
        passportNumber:
          type: string
          nullable: true
        travelDate:
          type: string
          format: date-time
          nullable: true
        returnDate:
          type: string
          format: date-time
          nullable: true
        purpose:
          type: string
          nullable: true
        price:
          type: number
          nullable: true
        isPaid:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
          nullable: true
        unreadCount:
          type: integer
      required:
        - id
        - userId
        - applicantName
        - destination
        - visaType
        - status
        - isPaid
        - createdAt
        - updatedAt
        - unreadCount

    CreateApplicationInput:
      type: object
      properties:
        applicantName:
          type: string
        destination:
          type: string
          enum: [usa, dubai, turkey, india]
        visaType:
          type: string
        passportNumber:
          type: string
        travelDate:
          type: string
          format: date-time
        returnDate:
          type: string
          format: date-time
        purpose:
          type: string
        notes:
          type: string
      required:
        - applicantName
        - destination
        - visaType

    UpdateApplicationInput:
      type: object
      properties:
        status:
          type: string
          enum: [draft, submitted, in_review, approved, rejected, appointment_scheduled]
        appointmentDate:
          type: string
          format: date-time
          nullable: true
        adminNotes:
          type: string
          nullable: true
        price:
          type: number
          nullable: true
        isPaid:
          type: boolean
        notes:
          type: string
          nullable: true

    ApplicationListResponse:
      type: object
      properties:
        applications:
          type: array
          items:
            $ref: "#/components/schemas/Application"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - applications
        - total
        - page
        - totalPages

    Message:
      type: object
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        userId:
          type: integer
        content:
          type: string
        isFromAdmin:
          type: boolean
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
        senderName:
          type: string
      required:
        - id
        - applicationId
        - userId
        - content
        - isFromAdmin
        - isRead
        - createdAt
        - senderName

    SendMessageInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    AdminStats:
      type: object
      properties:
        totalApplications:
          type: integer
        pendingReview:
          type: integer
        approvedThisMonth:
          type: integer
        totalClients:
          type: integer
        byDestination:
          type: object
          additionalProperties:
            type: integer
        byStatus:
          type: object
          additionalProperties:
            type: integer
        recentApplications:
          type: array
          items:
            $ref: "#/components/schemas/Application"
      required:
        - totalApplications
        - pendingReview
        - approvedThisMonth
        - totalClients
        - byDestination
        - byStatus
        - recentApplications