# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Joventy.cd — Visa Premium SaaS API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminStatsByDestination } from "./adminStatsByDestination";
import type { AdminStatsByStatus } from "./adminStatsByStatus";
import type { Application } from "./application";

export interface AdminStats {
  totalApplications: number;
  pendingReview: number;
  approvedThisMonth: number;
  totalClients: number;
  byDestination: AdminStatsByDestination;
  byStatus: AdminStatsByStatus;
  recentApplications: Application[];
}