# lib/api-zod/src/generated/types/applicationListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Joventy.cd — Visa Premium SaaS API
 * OpenAPI spec version: 0.1.0
 */
import type { Application } from "./application";

export interface ApplicationListResponse {
  applications: Application[];
  total: number;
  page: number;
  totalPages: number;
}