# lib/api-zod/src/generated/types/createApplicationInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Joventy.cd — Visa Premium SaaS API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateApplicationInputDestination } from "./createApplicationInputDestination";

export interface CreateApplicationInput {
  applicantName: string;
  destination: CreateApplicationInputDestination;
  visaType: string;
  passportNumber?: string;
  travelDate?: Date;
  returnDate?: Date;
  purpose?: string;
  notes?: string;
}